api update

+++ Mines/StartScene.cpp
@@ -31,7 +31,10 @@ void start_scene::on_start()
     rect_i const menuBounds {windowSize.Height, 0, windowSize.Width - windowSize.Height, windowSize.Height};
     _mainForm = std::make_shared<main_menu>(resGrp, menuBounds);
     _mainForm->BtnStart->Click.connect([&, windowSize](auto const&) {
-        _playField->start({_mainForm->SldWidth->Value, _mainForm->SldHeight->Value}, windowSize.Height, _mainForm->SldMines->Value);
+        _playField->start(
+            {static_cast<i32>(_mainForm->SldWidth->Value), static_cast<i32>(_mainForm->SldHeight->Value)},
+            windowSize.Height,
+            _mainForm->SldMines->Value);
     });
     _mainForm->BtnQuit->Click.connect([&](auto const&) {
         parent().pop_current_scene();
